requirements.txt
requests
google-generativeai
pandas
python-dotenv
apify-client
supabase
